# Addressable 8kb