--- a/src/Paprika.Tests/Store/AbandonedTests.cs
+++ b/src/Paprika.Tests/Store/AbandonedTests.cs
@@ -99,7 +99,7 @@ public class AbandonedTests : BasePageTests
         var keccaks = Initialize(accounts);
 
         // set big value
-        var accountValue = new byte[3000];
+        var accountValue = new byte[2900];
         new Random(17).NextBytes(accountValue);
 
         using var db = PagedDb.NativeMemoryDb(pageCount * Page.PageSize, HistoryDepth);